Make ZED to avoid replacing drives glitching for less than 5s
Sometimes a drive will glitch and disappear from the server for a
short time. ZFS will mark the drive as 'UNAVAILABLE' and then as
'ONLINE' again, but ZED will replace the drive with a spare either
way.

ZED goes through the events sequentially, so adding a simple "sleep"
won't solve this issue - we need to catch the event, reverify this and
silently drop it if we think it's just a glitch.

Implementation decisions:

 * we run `/sbin/zpool` because:

   - I don't want to learn how to use libzfs

   - libzfs does not have a stable API, it's supposed to be used
     internally for zfs developers.

 * we run `zpool list -H` because it has smaller output and it will
   fit into static 4096B buffer, making it easier to code.

Signed-off-by: Krzysztof Piecuch <kris@kentik.com>